Add files via upload
This project is based on a dataset that was sourced from kaggle. The goal was to display whether a job posting is fake or real using the elements of the description, such as easy-to-believe catchphrases, missing qualifications or absence of work experience. I have approached this with two methods. Logisitic Regression and Random Forest Classifier.